pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "apaudel96", email = "apaudel96@ymail.com"},
]
dependencies = [
    "bottle>=0.12.21",
    "python-dotenv>=0.20.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
